Tests show plenty of bugs to fix, but generally working.

+++ Movedata.h
@@ -3,24 +3,25 @@
 
 #include "config.h"
 #include "Point.h"
+#include <stdint.h>
 
 class Movedata
 {
 public:
   float feed;
 
-  unsigned long movesteps;
-  unsigned long axismovesteps[NUM_AXES];
+  uint32_t movesteps;
+  uint32_t axismovesteps[NUM_AXES];
   bool          axisdirs[NUM_AXES];
   int           leading_axis;
 
-  unsigned long startinterval;
-  unsigned long currentinterval;
-  unsigned long fullinterval;
-  unsigned long steps_to_accel;
-  unsigned long accel_until;
-  unsigned long decel_from;
-  unsigned long accel_inc;
+  uint32_t startinterval;
+  uint32_t currentinterval;
+  uint32_t fullinterval;
+  uint32_t steps_to_accel;
+  uint32_t accel_until;
+  uint32_t decel_from;
+  uint32_t accel_inc;
 
   Point         endpos;
   Point         startpos;
